modules/llm_interface.py
"""Module for interfacing with OpenAI LLMs and embeddings."""

import logging

from llama_index.embeddings.openai import OpenAIEmbedding
from llama_index.llms.openai import OpenAI

import config

logger = logging.getLogger(__name__)


def create_openai_embedding() -> OpenAIEmbedding:
    """Creates an OpenAI embedding model for vector representation.

    Returns:
        OpenAI embedding model.
    """
    openai_embedding = OpenAIEmbedding(
        model=config.EMBEDDING_MODEL_ID,
        api_key=config.OPENAI_API_KEY or None,
        api_base=config.OPENAI_BASE_URL or None,
    )
    logger.info(f"Created OpenAI embedding model: {config.EMBEDDING_MODEL_ID}")
    return openai_embedding


def create_openai_llm(
    temperature: float = config.TEMPERATURE,
    max_new_tokens: int = config.MAX_NEW_TOKENS,
    decoding_method: str = "sample",
    model_id: str | None = None,
) -> OpenAI:
    """Creates an OpenAI LLM for generating responses.

    Args:
        temperature: Temperature for controlling randomness in generation (0.0 to 1.0).
        max_new_tokens: Maximum number of new tokens to generate.
        decoding_method: Unused for OpenAI models; kept for interface compatibility.

    Returns:
        OpenAI LLM model.
    """
    _ = decoding_method

    resolved_model_id = model_id or config.LLM_MODEL_ID

    openai_llm = OpenAI(
        model=resolved_model_id,
        api_key=config.OPENAI_API_KEY or None,
        api_base=config.OPENAI_BASE_URL or None,
        temperature=temperature,
        max_tokens=max_new_tokens,
    )

    logger.info(f"Created OpenAI LLM model: {resolved_model_id}")
    return openai_llm


def change_llm_model(new_model_id: str) -> None:
    """Change the LLM model to use.

    Args:
        new_model_id: New LLM model ID to use.
    """
    global config
    config.LLM_MODEL_ID = new_model_id
    logger.info(f"Changed LLM model to: {new_model_id}")
