pyproject.toml
[project]
name = "llama-index-example-1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.12.0"
dependencies = [
    "llama-index-core==0.14.15",
    "llama-index-llms-openai==0.6.21",
    "llama-index-embeddings-openai==0.5.1",
    "llama-index-readers-web==0.5.6",
    "requests==2.32.2",
    "pydantic==2.12.5",
    "gradio==6.6.0",
    "python-dotenv==1.0.0",
]

[tool.ruff]
line-length = 120
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["modules", "config"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
[tool.pylint]
max-line-length = 120

[tool.pycodestyle]
max-line-length = 120
